Move compilation issue into its own module
