+++ rust/kcl-python-bindings/src/lib.rs
@@ -42,6 +42,7 @@ use serde::Serialize;
 use uuid::Uuid;
 
 use crate::bridge::bounding_box::BoundingBoxResponse;
+use crate::bridge::compilation_issue::CompilationIssue;
 use crate::bridge::physical_properties::PhysicalPropertiesRequest;
 use crate::bridge::physical_properties::PhysicalPropertiesResponse;
 use crate::bridge::sketch_constraints::KclErrorInfo;
@@ -262,40 +263,6 @@ async fn new_context_state(
     };
     let state = kcl_lib::ExecState::new(&ctx);
     Ok((ctx, state))
-}
-
-/// Wrapper for [kcl_lib::kcl_error::CompilationIssue].
-#[pyo3_stub_gen::derive::gen_stub_pyclass]
-#[pyclass(from_py_object)]
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct CompilationIssue {
-    inner: kcl_lib::CompilationIssue,
-}
-
-impl From<kcl_lib::kcl_error::CompilationIssue> for CompilationIssue {
-    fn from(value: kcl_lib::kcl_error::CompilationIssue) -> Self {
-        Self { inner: value }
-    }
-}
-
-#[pyo3_stub_gen::derive::gen_stub_pymethods]
-#[pymethods]
-impl CompilationIssue {
-    pub fn is_warning(&self) -> bool {
-        self.inner.severity.is_warning()
-    }
-
-    pub fn is_err(&self) -> bool {
-        self.inner.severity.is_err()
-    }
-
-    pub fn is_fatal(&self) -> bool {
-        self.inner.severity.is_fatal()
-    }
-
-    pub fn message(&self) -> &str {
-        &self.inner.message
-    }
 }
 
 /// Returned from execution functions.
